+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.0] - 2025-07-22
+
+### Added
+
+-   `UveDX::UveFileManager::createWADFile` function to create `.dat` assets file
+
+### Fixed
+
+-   Some clang-tidy hints
+
 ## [0.6.0] - 2025-06-11
 
 ### Added